requirements.txt
torch
torchvision
pytorch-ignite
imutils
dlib
kaggle
tensorboardX
tqdm
pillow